+++ index.js
@@ -1,8 +1,10 @@
 import { getInHypixel } from './functions.js';
-import { data } from './utilities/bots.js';    
+import { data } from './utilities/bots.js';  
+import { prefixData } from './utilities/prefix.js';  
 import './guild.js';
+import './utilities/best.js';
 
-import './utilities/cmdHelper.js'
+// import './utilities/cmdHelper.js'
 //! any misc problems (/ct dump and copy message here)
 //? ...
 
@@ -41,3 +43,20 @@ register('gameLoad', () => {
         data.firstInstall = true;
     };      
 });
+
+
+register('command', () => {
+    const bulletSpace = '   &bo&r  ';
+    ChatLib.chat(ChatLib.getChatBreak(' '))
+    ChatLib.chat(`&6[&r&3Better Shridge&6]&r &cHelpline -------------`);
+    ChatLib.chat(`${bulletSpace}&f/setbotprefix &3- Sets Bot Prefix &r| &eCurrent: &r['${prefixData.bot}&r']`)
+    ChatLib.chat(`${bulletSpace}&f/setguildprefix &3- Sets Guild Prefix &r| &eCurrent: &r['${prefixData.guild}&r']`)
+    ChatLib.chat(`${bulletSpace}&f/setarrowprefix &3- Sets Arrow Prefix &r| &eCurrent: &r['${prefixData.arrow}&r']`)
+    ChatLib.chat(`${bulletSpace}&f/resetprefix (&9bot&r | &9guild&r | &9arrow&r) &3- Resets all prefixes`)
+    ChatLib.chat(' ');
+    ChatLib.chat(`      &6<&3Guild Best System&6> &b------ &r(/guildbest | /gb)`)
+    ChatLib.chat(`${bulletSpace}&f/guildbest list &3- Shows all the names in the guild best list`)                
+    ChatLib.chat(`${bulletSpace}&f/guildbest (name) &3- Sets/Adds a name to the guild best list \n      &c(Do the command again to remove name)`)
+    ChatLib.chat(`${bulletSpace}&f/setbestcolor &3- Sets color for guild best list names \n      &eCurrent: &r['${prefixData.best}test name&r'] `)
+    ChatLib.chat(ChatLib.getChatBreak(' '))
+}).setName('bettershridge').setAliases('bshelp', 'bs');